android/deploy-api/requirements.txt
nltk==3.2.5
textblob==0.15.1
jsonschema==2.6.0
google-cloud-vision==0.30.0
pandas==0.21.0
